sync

+++ app/src/main/java/utm/ptm/mtransport/data/models/Trip.java
@@ -6,6 +6,6 @@ public class Trip {
     public List<Route> routes;
     public List<Stop> stops;
     public List<Way> ways;
-    public double time;
+    public float time;
     public float cost;
 }
